Small refactoring

+++ MainWindow.xaml.cs
@@ -47,7 +47,7 @@ namespace MatrixCalculator
 
                 for (int c = 0; c < m.GetLength(1); c++)
                 {
-                    dataTable.Columns.Add(new DataColumn("", typeof(int)));
+                    dataTable.Columns.Add(new DataColumn("", typeof(double)));
                 }
 
 
@@ -149,7 +149,7 @@ namespace MatrixCalculator
 
                 for (int c = 0; c < m.GetLength(1); c++)
                 {
-                    dataTable.Columns.Add(new DataColumn("",typeof(int)));
+                    dataTable.Columns.Add(new DataColumn("",typeof(double)));
                 }
